// packages/artnet-packets/lib/poll-reply/poll-reply.interface.ts
export type PollReplyPacketPayload = {
  ipAddress: number[];
  port: number;
  firmwareVersion: number;
  netSwitch: number;
  netSubSwitch: number;
  oem: number;
  ubeaVersion: number;
  status1: number;
  estaManufactorerCode: number;
  shortName: string;
  longName: string;
  nodeReport: string;
  numPorts: number;
  portTypes: number[];
  goodInput: number[];
  goodOutputA: number[];
  swIn: number[];
  swOut: number[];
  acnPriority: number;
  swMacro: number;
  swRemote: number;
  spare: number[];
  style: number;
  macAddress: number[];
  bindIp: number[];
  bindIndex: number;
  status2: number;
  goodOutputB: number;
  status3: number;
  defaultRespUID: number[];
  filler: number[];
};